Suppressed the SLF4J no providers warning
SLF4J comes in as a transitive dependency of org.beryx:text-io, and once on my classpath, insists that I use it as a logging framework by polluting my terminal output with a warning.
I tried to switch to SLF4J to no avail, and decided instead to wrap my existing Log4J logger with a partial copy of the org.slf4j.Logger interface. This suppresses the warning and satisfies any uses of that API on my classpath.